Feature/dk/content processor

+++ Backend/WriteAPI/BI/Services/Notes/FullNotePhotosCollectionHandlerCommand.cs
@@ -7,32 +7,34 @@ using BI.Helpers;
 using BI.Services.History;
 using BI.SignalR;
 using Common.DatabaseModels.Models.NoteContent.FileContent;
+using Common.DTO;
 using Common.DTO.Notes.FullNoteContent;
 using Domain.Commands.Files;
 using Domain.Commands.NoteInner.FileContent.Photos;
 using Domain.Queries.Permissions;
 using MediatR;
-using WriteContext.Repositories;
+using WriteContext.Repositories.Files;
 using WriteContext.Repositories.NoteContent;
 
 namespace BI.Services.Notes
 {
-    public class FullNotePhotosCollectionHandlerCommand :
-        IRequestHandler<RemovePhotosCollectionCommand, OperationResult<Unit>>,
+    public class FullNotePhotosCollectionHandlerCommand : FullNoteBaseCollection,
+        IRequestHandler<UnlinkPhotosCollectionCommand, OperationResult<Unit>>,
         IRequestHandler<UploadPhotosToCollectionCommand, OperationResult<List<PhotoNoteDTO>>>,
         IRequestHandler<RemovePhotoFromCollectionCommand, OperationResult<Unit>>,
         IRequestHandler<ChangePhotosCollectionRowCountCommand, OperationResult<Unit>>,
         IRequestHandler<ChangePhotosCollectionSizeCommand, OperationResult<Unit>>,
-        IRequestHandler<TransformToPhotosCollectionCommand, OperationResult<PhotosCollectionNoteDTO>>
+        IRequestHandler<TransformToPhotosCollectionCommand, OperationResult<PhotosCollectionNoteDTO>>,
+        IRequestHandler<UpdatePhotosContentsCommand, OperationResult<Unit>>
     {
 
         private readonly IMediator _mediator;
 
         private readonly BaseNoteContentRepository baseNoteContentRepository;
 
-        private readonly FileRepository fileRepository;
+        private readonly PhotosCollectionNoteRepository photosCollectionNoteRepository;
 
-        private readonly PhotosCollectionNoteRepository albumNoteRepository;
+        private readonly PhotoNoteAppFileRepository photoNoteAppFileRepository;
 
         private readonly HistoryCacheService historyCacheService;
 
@@ -42,21 +44,22 @@ namespace BI.Services.Notes
             IMediator _mediator,
             BaseNoteContentRepository baseNoteContentRepository,
             FileRepository fileRepository,
-            PhotosCollectionNoteRepository albumNoteRepository,
+            AppFileUploadInfoRepository appFileUploadInfoRepository,
+            PhotosCollectionNoteRepository photosCollectionNoteRepository,
+            PhotoNoteAppFileRepository photoNoteAppFileRepository,
             HistoryCacheService historyCacheService,
-            AppSignalRService appSignalRService)
+            AppSignalRService appSignalRService) : base(appFileUploadInfoRepository, fileRepository)
         {
             this._mediator = _mediator;
             this.baseNoteContentRepository = baseNoteContentRepository;
-            this.fileRepository = fileRepository;
-            this.albumNoteRepository = albumNoteRepository;
+            this.photosCollectionNoteRepository = photosCollectionNoteRepository;
+            this.photoNoteAppFileRepository = photoNoteAppFileRepository;
             this.historyCacheService = historyCacheService;
             this.appSignalRService = appSignalRService;
         }
 
 
-        // TODO REMOVE WITHOUT ORDERING
-        public async Task<OperationResult<Unit>> Handle(RemovePhotosCollectionCommand request, CancellationToken cancellationToken)
+        public async Task<OperationResult<Unit>> Handle(UnlinkPhotosCollectionCommand request, CancellationToken cancellationToken)
         {
             var command = new GetUserPermissionsForNoteQuery(request.NoteId, request.Email);
             var permissions = await _mediator.Send(command);
@@ -64,42 +67,36 @@ namespace BI.Services.Notes
 
             if (permissions.CanWrite)
             {
+                var photos = await photoNoteAppFileRepository.GetWhereAsync(x => x.PhotosCollectionNoteId == request.ContentId);
+                var ids = photos.Select(x => x.AppFileId).ToArray();
 
-                var contents = await baseNoteContentRepository.GetAllContentByNoteIdOrdered(note.Id);
-                var contentForRemove = contents.FirstOrDefault(x => x.Id == request.ContentId) as PhotosCollectionNote;
-                contents.Remove(contentForRemove);
+                await MarkAsUnlinked(ids);
 
-                var orders = Enumerable.Range(1, contents.Count);
-                contents = contents.Zip(orders, (content, order) =>
-                {
-                    content.Order = order;
-                    content.UpdatedAt = DateTimeOffset.Now;
-                    return content;
-                }).ToList();
+                return new OperationResult<Unit>(success: true, Unit.Value);
+            }
 
+            return new OperationResult<Unit>().SetNoPermissions();
+        }
 
-                using var transaction = await baseNoteContentRepository.context.Database.BeginTransactionAsync();
+        public async Task<OperationResult<Unit>> Handle(RemovePhotoFromCollectionCommand request, CancellationToken cancellationToken)
+        {
+            var command = new GetUserPermissionsForNoteQuery(request.NoteId, request.Email);
+            var permissions = await _mediator.Send(command);
+            var note = permissions.Note;
 
-                try
-                {
-                    await baseNoteContentRepository.RemoveAsync(contentForRemove);
-                    await baseNoteContentRepository.UpdateRangeAsync(contents);
+            if (permissions.CanWrite)
+            {
+                var collection = await photosCollectionNoteRepository.GetOneIncludePhotoNoteAppFiles(request.ContentId);
+                collection.PhotoNoteAppFiles = collection.PhotoNoteAppFiles.Where(x => x.AppFileId != request.PhotoId).ToList();
+                collection.UpdatedAt = DateTimeOffset.Now;
 
-                    await transaction.CommitAsync();
+                await photosCollectionNoteRepository.UpdateAsync(collection);
+                await MarkAsUnlinked(request.PhotoId);
 
-                    await _mediator.Send(new RemoveFilesCommand(permissions.User.Id.ToString(), contentForRemove.Photos));
+                historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
+                await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
 
-                    historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
-
-                    await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
-
-                    return new OperationResult<Unit>(success: true, Unit.Value);
-                }
-                catch (Exception e)
-                {
-                    await transaction.RollbackAsync();
-                    Console.WriteLine(e);
-                }
+                return new OperationResult<Unit>(success: true, Unit.Value);
             }
 
             return new OperationResult<Unit>().SetNoPermissions();
@@ -112,10 +109,10 @@ namespace BI.Services.Notes
 
             if (permissions.CanWrite)
             {
-                var album = await baseNoteContentRepository.GetContentById<PhotosCollectionNote>(request.ContentId);
-                album.CountInRow = request.Count;
-                album.UpdatedAt = DateTimeOffset.Now;
-                await baseNoteContentRepository.UpdateAsync(album);
+                var collection = await photosCollectionNoteRepository.FirstOrDefaultAsync(x => x.Id == request.ContentId);
+                collection.CountInRow = request.Count;
+                collection.UpdatedAt = DateTimeOffset.Now;
+                await photosCollectionNoteRepository.UpdateAsync(collection);
 
                 historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
                 await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
@@ -133,47 +130,13 @@ namespace BI.Services.Notes
 
             if (permissions.CanWrite)
             {
-                var album = await baseNoteContentRepository.GetContentById<PhotosCollectionNote>(request.ContentId);
-                album.Height = request.Height;
-                album.Width = request.Width;
-                album.UpdatedAt = DateTimeOffset.Now;
-                await baseNoteContentRepository.UpdateAsync(album);
+                var collection = await photosCollectionNoteRepository.FirstOrDefaultAsync(x => x.Id == request.ContentId);
+                collection.Height = request.Height;
+                collection.Width = request.Width;
+                collection.UpdatedAt = DateTimeOffset.Now;
+                await baseNoteContentRepository.UpdateAsync(collection);
 
                 historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
-                await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
-
-                return new OperationResult<Unit>(success: true, Unit.Value);
-            }
-
-            return new OperationResult<Unit>().SetNoPermissions();
-        }
-
-        public async Task<OperationResult<Unit>> Handle(RemovePhotoFromCollectionCommand request, CancellationToken cancellationToken)
-        {
-            var command = new GetUserPermissionsForNoteQuery(request.NoteId, request.Email);
-            var permissions = await _mediator.Send(command);
-            var note = permissions.Note;
-
-            if (permissions.CanWrite)
-            {
-                var album = await baseNoteContentRepository.GetContentById<PhotosCollectionNote>(request.ContentId);
-                var photoForRemove = album.Photos.First(x => x.Id == request.PhotoId);
-                album.Photos.Remove(photoForRemove);
-                album.UpdatedAt = DateTimeOffset.Now;
-
-                if (album.Photos.Count == 0)
-                {
-                    var resp = await _mediator.Send(new RemovePhotosCollectionCommand(note.Id, album.Id, request.Email));
-                }
-                else
-                {
-                    await baseNoteContentRepository.UpdateAsync(album);
-                }
-
-                await _mediator.Send(new RemoveFilesCommand(permissions.User.Id.ToString(), photoForRemove));
-
-                historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
-
                 await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
 
                 return new OperationResult<Unit>(success: true, Unit.Value);
@@ -210,18 +173,18 @@ namespace BI.Services.Notes
                 }
 
                 // UPDATING
-                var album = await baseNoteContentRepository.GetContentById<PhotosCollectionNote>(request.ContentId);
+                var collection = await photosCollectionNoteRepository.GetOneIncludePhotos(request.ContentId);
 
                 using var transaction = await baseNoteContentRepository.context.Database.BeginTransactionAsync();
 
                 try
                 {
-                    await fileRepository.AddRangeAsync(dbFiles);
+                    collection.Photos.AddRange(dbFiles);
+                    collection.UpdatedAt = DateTimeOffset.Now;
 
-                    album.Photos.AddRange(dbFiles);
-                    album.UpdatedAt = DateTimeOffset.Now;
+                    await photosCollectionNoteRepository.UpdateAsync(collection);
 
-                    await albumNoteRepository.UpdateAsync(album);
+                    await MarkAsLinked(dbFiles);
 
                     await transaction.CommitAsync();
 
@@ -252,13 +215,19 @@ namespace BI.Services.Notes
             {
                 var contentForRemove = await baseNoteContentRepository.FirstOrDefaultAsync(x => x.Id == request.ContentId);
 
+                if (contentForRemove == null)
+                {
+                    throw new Exception("Content not found");
+                }
+
+
                 using var transaction = await baseNoteContentRepository.context.Database.BeginTransactionAsync();
 
                 try
                 {
                     await baseNoteContentRepository.RemoveAsync(contentForRemove);
 
-                    var albumNote = new PhotosCollectionNote()
+                    var photosCollection = new PhotosCollectionNote()
                     {
                         NoteId = request.NoteId,
                         Order = contentForRemove.Order,
@@ -267,12 +236,12 @@ namespace BI.Services.Notes
                         Height = "auto",
                     };
 
-                    await albumNoteRepository.AddAsync(albumNote);
+                    await photosCollectionNoteRepository.AddAsync(photosCollection);
 
                     await transaction.CommitAsync();
 
-                    var result = new PhotosCollectionNoteDTO(null, albumNote.Width, albumNote.Height,
-                        albumNote.Id, albumNote.CountInRow, albumNote.UpdatedAt);
+                    var result = new PhotosCollectionNoteDTO(null, photosCollection.Name, photosCollection.Width, photosCollection.Height,
+                        photosCollection.Id, photosCollection.Order, photosCollection.CountInRow, photosCollection.UpdatedAt);
 
                     historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
                     await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
@@ -287,6 +256,76 @@ namespace BI.Services.Notes
             }
 
             return new OperationResult<PhotosCollectionNoteDTO>().SetNoPermissions();
+        }
+
+        public async Task<OperationResult<Unit>> Handle(UpdatePhotosContentsCommand request, CancellationToken cancellationToken)
+        {
+            var command = new GetUserPermissionsForNoteQuery(request.NoteId, request.Email);
+            var permissions = await _mediator.Send(command);
+
+            if (permissions.CanWrite)
+            {
+                if (request.Photos.Count == 1)
+                {
+                    await UpdateOne(request.Photos.First());
+                }
+                else
+                {
+                    await UpdateMany(request.Photos);
+                }
+
+                historyCacheService.UpdateNote(permissions.Note.Id, permissions.User.Id, permissions.Author.Email);
+                await appSignalRService.UpdateContent(request.NoteId, permissions.User.Email);
+
+                // TODO DEADLOCK
+                return new OperationResult<Unit>(success: true, Unit.Value);
+            }
+
+            return new OperationResult<Unit>().SetNoPermissions();
+        }
+
+        private async Task UpdateMany(List<PhotosCollectionNoteDTO> entities)
+        {
+            foreach (var entity in entities)
+            {
+                await UpdateOne(entity);
+            }
+        }
+
+        private async Task UpdateOne(PhotosCollectionNoteDTO entity)
+        {
+            var entityForUpdate = await photosCollectionNoteRepository.GetOneIncludePhotoNoteAppFiles(entity.Id);
+            if (entityForUpdate != null)
+            {
+                entityForUpdate.UpdatedAt = DateTimeOffset.Now;
+                entityForUpdate.Name = entity.Name;
+                entityForUpdate.Width = entity.Width;
+                entityForUpdate.Height = entity.Height;
+                entityForUpdate.CountInRow = entity.CountInRow;
+
+                var databaseFileIds = entityForUpdate.PhotoNoteAppFiles.Select(x => x.AppFileId);
+                var entityFileIds = entity.Photos.Select(x => x.FileId);
+
+                var idsForDelete = databaseFileIds.Except(entityFileIds);
+                var idsForAdd = entityFileIds.Except(databaseFileIds);
+
+                if (idsForDelete.Any() || idsForAdd.Any())
+                {
+                    entityForUpdate.PhotoNoteAppFiles = entity.Photos.Select(x =>
+                        new PhotoNoteAppFile { AppFileId = x.FileId, PhotosCollectionNoteId = entityForUpdate.Id }).ToList();
+                }
+
+                if (idsForDelete.Any())
+                {
+                    await MarkAsUnlinked(idsForDelete.ToArray());
+                }
+
+                if (idsForAdd.Any())
+                {
+                    await MarkAsLinked(idsForAdd.ToArray());
+                }
+            }
+            await photosCollectionNoteRepository.UpdateAsync(entityForUpdate);
         }
     }
 }
